Distance should be a decimal

+++ main/models.py
@@ -44,7 +44,7 @@ class Race(models.Model):
     name = models.CharField(max_length=254)
     start_date = models.DateField()
     end_date = models.DateField()
-    distance = models.IntegerField()
+    distance = models.DecimalField(max_digits=5, decimal_places=2)
     submissions_close = models.DateField()
     DISTANCE_UNIT_CHOICES = [
         ("M", "Miles"),
